double check for sufficient memory to run ZRAM

+++ functions/zram.bash
@@ -1,12 +1,18 @@
 #!/usr/bin/env bash
 
 init_zram_mounts() {
-  local introtext="You are about to activate the ZRAM feature.\\nBe aware this is a dangerous operation to apply to your system. Use is at your own risk of data loss.\\nPlease check out the \"ZRAM status\" thread at https://community.openhab.org/t/zram-status/80996 before proceeding."
+  local introtext="You are about to activate the ZRAM feature.\\nBe aware you do this at your own risk of data loss.\\nPlease check out the \"ZRAM status\" thread at https://community.openhab.org/t/zram-status/80996 before proceeding."
+  local text_lowmem="Your system has less than 1 GB of RAM. It is definitely NOT recommended to run ZRAM (AND openHAB) on your box. If you proceed now you will do so at your own risk !"
   if [ "$1" == "install" ]; then
     if [ -z "$UNATTENDED" ]; then
-      # ... display warn disclaimer...
-      # point to ZRAM status thread on forum
-      if ! (whiptail --title "Install ZRAM, Continue?" --yes-button "Continue" --no-button "Back" --yesno "$introtext" 15 80); then echo "CANCELED"; return 0; fi
+      # display warn disclaimer and point to ZRAM status thread on forum
+      if ! (whiptail --title "Install ZRAM, Continue?" --yes-button "Continue" --no-button "Back" --yesno "$introtext" 15 80) then echo "CANCELED"; return 0; fi
+      # double check if there's enough RAM to run ZRAM
+      lowmemory=false
+      if has_lowmem; then
+        lowmemory=true
+        if ! (whiptail --title "WARNING, Continue?" --yes-button "REALLY Continue" --no-button "Step Back" --yesno --defaultno "$text_lowmem" 15 80) then echo "CANCELED"; return 0; fi
+      fi
     fi
 
     local ZRAMGIT=https://github.com/mstormi/openhabian-zram
@@ -29,13 +35,13 @@ init_zram_mounts() {
 
 zram_setup() {
   if is_arm; then
-    if ! is_pione && ! is_cmone && ! is_pizero && ! is_pizerow; then
+    if ! has_lowmem && ! is_pione && ! is_cmone && ! is_pizero && ! is_pizerow; then
       cond_redirect systemctl stop openhab2
       echo -n "$(timestamp) [openHABian] Installing ZRAM ..."
       init_zram_mounts install
       cond_redirect systemctl start openhab2
     else
-      echo -n "$(timestamp) [openHABian] Skipping ZRAM install on hardware without enough memory."
+      echo -n "$(timestamp) [openHABian] Skipping ZRAM install on ARM hardware without enough memory."
     fi
   else
     echo -n "$(timestamp) [openHABian] Skipping ZRAM install on non-ARM hardware."
